requirements.txt
joblib==1.3.2
matplotlib==3.5.2
networkx==3.1
numpy==1.24.4
pandas==2.2.2
pyserial==3.5
PyYAML==6.0.2
seaborn==0.13.2
sphinx_rtd_theme==2.0.0
tqdm==4.66.1
